Refactor: move combined SARIF debug artifact logic to `debug-artifact`
