+++ lib/debug-artifacts.js
@@ -27,6 +27,7 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.sanitizeArifactName = sanitizeArifactName;
+exports.uploadCombinedSarifArtifacts = uploadCombinedSarifArtifacts;
 exports.uploadAllAvailableDebugArtifacts = uploadAllAvailableDebugArtifacts;
 exports.uploadDebugArtifacts = uploadDebugArtifacts;
 const fs = __importStar(require("fs"));
@@ -42,6 +43,33 @@ const environment_1 = require("./environment");
 const util_1 = require("./util");
 function sanitizeArifactName(name) {
     return name.replace(/[^a-zA-Z0-9_\\-]+/g, "");
+}
+/**
+ * Upload Actions SARIF artifacts for debugging when CODEQL_ACTION_DEBUG_COMBINED_SARIF
+ * environment variable is set
+ */
+async function uploadCombinedSarifArtifacts() {
+    const tempDir = (0, actions_util_1.getTemporaryDirectory)();
+    // Upload Actions SARIF artifacts for debugging when environment variable is set
+    if (process.env["CODEQL_ACTION_DEBUG_COMBINED_SARIF"] === "true") {
+        core.info("Uploading available combined SARIF files as Actions debugging artifact...");
+        const baseTempDir = path.resolve(tempDir, "combined-sarif");
+        const toUpload = [];
+        if (fs.existsSync(baseTempDir)) {
+            const outputDirs = fs.readdirSync(baseTempDir);
+            for (const outputDir of outputDirs) {
+                const sarifFiles = fs
+                    .readdirSync(path.resolve(baseTempDir, outputDir))
+                    .filter((f) => f.endsWith(".sarif"));
+                for (const sarifFile of sarifFiles) {
+                    toUpload.push(path.resolve(baseTempDir, outputDir, sarifFile));
+                }
+            }
+        }
+        if (toUpload.length > 0) {
+            await uploadDebugArtifacts(toUpload, baseTempDir, "upload-debug-artifacts");
+        }
+    }
 }
 async function uploadAllAvailableDebugArtifacts(config, logger) {
     const filesToUpload = [];
